+++ TableViewKitTests/ArrayChangesUtils.swift
@@ -3,32 +3,26 @@ import Foundation
 
 extension ArrayChanges: Equatable {
     public static func == (lhs: ArrayChanges<Element>, rhs: ArrayChanges<Element>) -> Bool {
+        guard
+            let lhs = lhs as? ArrayChanges<Int>,
+            let rhs = rhs as? ArrayChanges<Int> else { return false }
         switch (lhs, rhs) {
         case (let .inserts(indexesLhs, elementsLhs),
               let .inserts(indexesRhs, elementsRhs)),
              (let .deletes(indexesLhs, elementsLhs),
               let .deletes(indexesRhs, elementsRhs)):
-            if let elementsLhs = elementsLhs as? [Int],
-                let elementsRhs = elementsRhs as? [Int] {
-                return indexesLhs == indexesRhs && elementsLhs == elementsRhs
-            } else {
-                return false
-            }
+            return indexesLhs == indexesRhs && elementsLhs == elementsRhs
         case (let .updates(indexesLhs),
               let .updates(indexesRhs)):
             return indexesLhs == indexesRhs
         case (let .moves(indexesLhs),
               let .moves(indexesRhs)):
-            guard indexesLhs.count == indexesRhs.count else {
-                return false
-            }
-            return zip(indexesLhs, indexesRhs).reduce(true, { (carry, arg) -> Bool in
-                let ((fromLhs, toLhs), (fromRhs, toRhs)) = arg
-                return carry && (fromLhs == fromRhs && toLhs == toRhs)
-            })
-        case (.beginUpdates, .beginUpdates),
-             (.endUpdates, .endUpdates):
-            return true
+            return indexesLhs.elementsEqual(indexesRhs, by: { $0.0 == $1.0 && $0.1 == $1.1 })
+        case (let .beginUpdates(fromElementsLhs, toElementsLhs),
+              let .beginUpdates(fromElementsRhs, toElementsRhs)),
+             (let .endUpdates(fromElementsLhs, toElementsLhs),
+              let .endUpdates(fromElementsRhs, toElementsRhs)):
+            return fromElementsLhs == fromElementsRhs && toElementsLhs == toElementsRhs
         default:
             return false
         }