+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ---
 ## [Unreleased]
 
+- Update to Swift 4.2 ([PR#42](https://github.com/odigeoteam/TableViewKit/pull/42))
 - Improved initial slicing for diff algorithm ([PR#39](https://github.com/odigeoteam/TableViewKit/pull/39))
 - Fixed ObservableArray setter ([PR#38](https://github.com/odigeoteam/TableViewKit/pull/38))
 - Improved/Fixed examples ([PR#34](https://github.com/odigeoteam/TableViewKit/pull/34), [PR#35](https://github.com/odigeoteam/TableViewKit/pull/35))
@@ -37,4 +38,3 @@ All notable changes to this project will be documented in this file.
 - Method `Item.select(in:animated:scrollPosition:)` has been renamed to `Item.select(animated:scrollPosition:)`
 - Method `Item.deselect(in:animated:)` has been renamed to `Item.deselect(animated:)`
 - Method `Section.index(in:)` has been renamed to `Section.index`
-